Refactor, move location to separate route, rename Location type to Coordinates
