+++ server/src/routes/users.ts
@@ -1,4 +1,3 @@
-import axios from 'axios';
 import express from 'express';
 import asyncHandler from 'express-async-handler';
 import { AppError } from '../errors';
@@ -199,39 +198,6 @@ router.get(
 		if (!req.params.id) throw new AppError(`Please, provide user id`, 400);
 		const userPhotos = await getUserPhotosById(req.params.id);
 		res.status(200).json(userPhotos);
-	})
-);
-
-router.put(
-	'/:id/location',
-	sessionExtractor,
-	asyncHandler(async (req: CustomRequest, res) => {
-		if (!req.session || !req.session.userId || req.session.userId !== req.params.id) {
-			throw new AppError(`No rights to update profile data`, 400);
-		}
-		// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-		const coordinates = req.body.coordinates;
-		// if (!coordinates) {
-		// 	console.log('ip ', req.headers['x-forwarded-for'] || req.socket.remoteAddress || null);
-		// }
-		// console.log(coordinates.latitude);
-		let result;
-		try {
-			const params = {
-				access_key: process.env.API_KEY,
-				query: `${coordinates[0]}, ${coordinates[1]}`
-			};
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call
-			// console.log(coordinates[0], coordinates[1]);
-			result = await axios.get(`http://api.positionstack.com/v1/reverse`, { params });
-		} catch (err) {
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-call
-
-			if (axios.isAxiosError(err)) console.log('Response err: ', err.response?.data);
-		}
-
-		// eslint-disable-next-line @typescript-eslint/no-unsafe-return
-		res.status(201).json(result?.data?.data[0]);
 	})
 );
