+++ server/src/services/users.ts
@@ -5,7 +5,7 @@ import { addUpdateEmailRequest, findUpdateEmailRequestByUserId, removeUpdateEmai
 //prettier-ignore
 import { addNewUser, findUserByActivationCode, setUserAsActive, findUserByEmail, updateUserPassword, updateUserEmail, getPasswordHash, isUserById } from '../repositories/userRepository';
 import { updateSessionEmailByUserId } from '../repositories/sessionRepository';
-import { EmailUpdateRequest, NewUser, PasswordResetRequest, Photo, User, Location } from '../types';
+import { EmailUpdateRequest, NewUser, PasswordResetRequest, Photo, User, Coordinates } from '../types';
 import { sendMail } from '../utils/mailer';
 import { AppError } from '../errors';
 import bcrypt from 'bcrypt';
@@ -19,8 +19,8 @@ export const createHashedPassword = async (passwordPlain: string): Promise<strin
 	return await bcrypt.hash(passwordPlain, saltRounds);
 };
 
-export const requestLocationByIp = async (ipAddress: string | undefined): Promise<Location> => {
-	const defaultCoordinates: Location = { lat: 60.16678195339881, lon: 24.941711425781254 }; //Hki city center as we are hki startup hehe
+export const requestCoordinatesByIp = async (ipAddress: string | undefined): Promise<Coordinates> => {
+	const defaultCoordinates: Coordinates = { lat: 60.16678195339881, lon: 24.941711425781254 }; //Hki city center as we are hki startup hehe
 	if (!ipAddress) return defaultCoordinates;
 
 	try {
@@ -54,7 +54,7 @@ export const createNewUser = async (newUser: NewUser, ipAddress: string | undefi
 	const passwordHash = await createHashedPassword(newUser.passwordPlain);
 	const activationCode = crypto.randomBytes(20).toString('hex');
 
-	const coordinates = await requestLocationByIp(ipAddress);
+	const coordinates = await requestCoordinatesByIp(ipAddress);
 	return addNewUser({ ...newUser, passwordHash, activationCode, lat: coordinates.lat, lon: coordinates.lon });
 };
