+++ server/src/types.ts
@@ -7,7 +7,7 @@ export type BaseUser = {
 	lastname: string;
 };
 
-export type Location = {
+export type Coordinates = {
 	lat: number;
 	lon: number;
 };
@@ -23,7 +23,7 @@ export type UserData = {
 	orientation: string | undefined;
 	bio: string | undefined;
 	tags: string[] | undefined;
-	coordinates: Location;
+	coordinates: Coordinates;
 	location: string;
 };
 
@@ -35,7 +35,7 @@ export type UpdateUserProfile = {
 	orientation: Orientation;
 	bio: string;
 	tags: string[];
-	coordinates: Location;
+	coordinates: Coordinates;
 	location: string;
 };
 
@@ -43,7 +43,7 @@ export type Gender = 'male' | 'female';
 
 export type Orientation = 'straight' | 'gay' | 'bi';
 
-export type User = BaseUser & { id: string; passwordHash: string; createdAt: Date; isActive: boolean; activationCode: string; coordinates: Location, location: string };
+export type User = BaseUser & { id: string; passwordHash: string; createdAt: Date; isActive: boolean; activationCode: string; coordinates: Coordinates, location: string };
 
 export type NewUserWithHashedPwd = BaseUser & { passwordHash: string; activationCode: string; lat: number; lon: number };
